embot::os is now more clearly separated from embot::hw. see details below.
embot::os ->

- reduced dependencies from embot::hw. now only embot_os.cpp and embot_hw.h are needed so that
  funtion embot::hw::init() can be called inside embot::os::init(). all the types required by
  hw drivers are inside embot_hw_type.h

  with this change we can now use the embot::os also on boards where the stm32hal is not running,
  and there is a different hardware abstraction layer, such as in the ems / mc4plus / mc2plus.

- added extra functions in embot::os::rtos such as messagequeue_size(), scheduler_lock(),
  scheduler_unlock(), memory_new(), memory_delete(), memory_realloc().

  these are required if we want to run embot::os onto ems / mc4plus / mc2plus.

- moved the default tick time function inside stm32hal.c, so that it is this library which decides
  how to get the 1 ms tick when the one passed at its initialization  is not valid.
  it was necessary to recompile all the stm32halxxxx.lib libraries.

- adjusted some projects (stm32hal* and some board's ones).
