+++ emBODY/eBcode/arch-arm/board/rfe/application/src/usb_user/embot_hw_usb.h
@@ -23,7 +23,7 @@
 #define _EMBOT_HW_USB_H_
 
 #include "embot_core.h"
-#include "embot_hw.h"
+#include "embot_hw_types.h"
