+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_ads122c04.h
@@ -11,7 +11,7 @@
 #define _EMBOT_HW_ADS122C04_H_
 
 #include "embot_core.h"
-#include "embot_hw.h"
+#include "embot_hw_types.h"
 
 
 #include "embot_hw_i2c.h"
