+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_button.h
@@ -23,7 +23,7 @@
 #define _EMBOT_HW_BUTTON_H_
 
 #include "embot_core.h"
-#include "embot_hw.h"
+#include "embot_hw_types.h"
