+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_gpio.h
@@ -23,8 +23,7 @@
 #define _EMBOT_HW_GPIO_H_
 
 #include "embot_core.h"
-#include "embot_hw.h"
-
+#include "embot_hw_types.h"
 
 namespace embot { namespace hw { namespace gpio {
