+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_led.h
@@ -24,7 +24,7 @@
 
 
 #include "embot_core.h"
-#include "embot_hw.h"
+#include "embot_hw_types.h"
 
 
 namespace embot { namespace hw { namespace led {
