+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_si7051.h
@@ -23,7 +23,7 @@
 #define _EMBOT_HW_SI7051_H_
 
 #include "embot_core.h"
-#include "embot_hw.h"
+#include "embot_hw_types.h"
 
 
 #include "embot_hw_i2c.h"
