+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_sys.h
@@ -23,7 +23,7 @@
 #define _EMBOT_HW_SYS_H_
 
 #include "embot_core.h"
-#include "embot_hw.h"
+#include "embot_hw_types.h"
 #include <string>
