+++ emBODY/eBcode/arch-arm/libs/lowlevel/stm32hal/proj/stm32hal.l4.uvoptx
@@ -1426,7 +1426,7 @@
       <GroupNumber>1</GroupNumber>
       <FileNumber>1</FileNumber>
       <FileType>1</FileType>
-      <tvExp>1</tvExp>
+      <tvExp>0</tvExp>
       <tvExpOptDlg>0</tvExpOptDlg>
       <bDave2>0</bDave2>
       <PathWithFileName>..\src\stm32hal\stm32hal.c</PathWithFileName>
